requirements.txt
streamlit==1.39.0
tensorflow==2.15.0
pillow==10.1.0
numpy==1.24.3
